requirements.txt
SpeechRecognition
pyaudio
keyboard
pyttsx3
pyautogui
Pillow
pyinstaller